remove redundant invoker module and add revoke method instead
