+++ lib/jets.rb
@@ -29,9 +29,7 @@ module Jets
   autoload :Db, 'jets/db'
   autoload :Dotenv, 'jets/dotenv'
   autoload :Erb, "jets/erb"
-  autoload :Invoker, 'jets/invoker'
   autoload :Generator, "jets/generator"
-  autoload :Degenerator, "jets/degenerator"
   autoload :Inflections, "jets/inflections"
   autoload :IO, "jets/io"
   autoload :Job, 'jets/job'
